Extract `on_resolution` and additional `redeem_shares` tests
