+++ zrml/prediction-markets/src/tests/mod.rs
@@ -24,12 +24,14 @@ mod approve_market;
 mod buy_complete_set;
 mod close_trusted_market;
 mod create_market;
+mod create_market_and_deploy_pool;
 mod dispute;
 mod dispute_early_close;
 mod edit_market;
 mod manually_close_market;
 mod on_initialize;
 mod on_market_close;
+mod on_resolution;
 mod redeem_shares;
 mod reject_early_close;
 mod reject_market;
